Added some features
Relative path behavior is a bit more defined.
Also added a flag to force absolute paths in exporting the checksum file.
The path flag's argument is now marked optional, to better align with the description.
Invalid lines note line number as the program is running, just in case.
Improper lines have an extra check: if the "hash" portion is not hex, then mark as bad format.
The checksum list sorting method should work properly now after a bit of testing.
Caught exceptions no longer print out the full stack trace since it is likely not to be relevant.
Changed version to something different so to prepare for another pre-release.